backend/pyproject.toml
[project]
name = "codebay-final-project"
version = "0.1.0"
description = ""
authors = [
    {name = "luis-barbara",email = "luis.carlos.barbara@gmail.com"}
]

requires-python = ">=3.12,<3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.68.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-allauth (>=65.9.0,<66.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "social-auth-app-django (>=5.4.3,<6.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "google-api-python-client (>=2.171.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "boto3 (>=1.38.31,<2.0.0)",
    "stripe (>=12.2.0,<13.0.0)",
    "stream-chat (>=4.24.0,<5.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)", 
]

[tool.poetry]
packages = [{include = "codebay"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
